+++ tests/testthat/test-myip.R
@@ -1,6 +1,6 @@
-context("basic functionality")
+context("Multi-service external lookups work")
 test_that("we can do something", {
 
-  #expect_that(some_function(), is_a("data.frame"))
+  expect_is(myip("all"), "list")
 
 })